Don't assume a default of 'false' for TwoStatePreference
TwoStatePreference would treat a default value of false as having no
effect since it matched the initial value of the internal field. This
caused asymmetric behavior around persistence and listener
notification when the default was true (persisted) vs. false (not
persisted).

Standard behavior for most Preferences is to persist the default
value. Make TwoStatePreference (CheckBoxPreference and
SwitchPreference) follow this pattern.

Bug 5722690

Change-Id: I8b3cf55efc5dfd573b511913b48ced017f0432d8