Improve systrace tag propagation
The native side of the tracing code latches a copy of the tags from a
system property on first use.  The Java-side tracing code latches a
copy of the native's copy during class init.  The tracing code is
preloaded by the zygote, which means we get the flags during zygote
init and don't update them when we launch a new app.

This changes the Java sources to also defer initialization until
first use, so that newly-launched apps will use the current value
of the system property.

Bug 7323431

Change-Id: I7db048ec54345ae9565088a35c2e2b4c82f993fd