Set wallpaper bitmap reference to null before reloading
Give the GC a chance to collect the current bitmap if it needs to,
as it allocates memory for the next one. This helps avoid OOM situations
that can sometimes occur in extreme circumstances (huge bitmaps)

Also set the default_wallpaper to the right default size.

Issue #7352961 Wallpaper edge is cut-off while scrolling through home page

Change-Id: If76b55061d04b29af7f66a6162e307b8b53bf4ae